Extract `make` into `make.py`
