+++ src/ninetoothed/__init__.py
@@ -1,4 +1,5 @@
-from ninetoothed.jit import jit, make
+from ninetoothed.jit import jit
+from ninetoothed.make import make
 from ninetoothed.symbol import Symbol
 from ninetoothed.tensor import Tensor
